Fixed Out-of-bounds Memory Access
There was a deep copy problem between the parameter class on the
host and the device. Passing around pointers to structures in
kernels is frowned upon anyway so the deep copy problem was
resolved by passing the parameter struct by value instead and
handling the one pointer (delta) separately as an argument.