Linc is now working
Memory access error bug was causing undefined behavior, likely the
culprit since the prior version of linc (in all versions
currently) still alive it was in fact a present error, though I
did not bother to see if it was the only error. The current working
version of linc uses thrust and i'm biased towards keeping it.

An extensive test of linc was also added, randomly generated
vectors and constants are used, computed in linear time on the cpu
and then tested against linc. If every computation is accurate to
the first place digit then the test is considered a success. A more
robust (relative error based) test will be implemented in the
future.