Restructuring
Not much has changed within the actual code with exception of
severe refactoring. From now on the design pattern is going to
emphasize using thrust whenever it makes sense to do so, or
whenever an opportunity to do so becomes apparent. A unit testing
paradigm has been implemented and macros are handled in a much
neater fashion.